+++ 1.21.1/src/main/java/com/torr/materia/ModCreativeTabs.java
@@ -307,6 +307,8 @@ public class ModCreativeTabs {
                 || path.contains("frame")
                 || path.contains("tile")
                 || path.endsWith("_block")
+                || path.endsWith("_column")
+                || path.endsWith("_column_capital")
                 || path.equals("salt_block");
     }